Added User roles for goals

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const express = require('express');
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel');
-const { error } = require('console');
+const User = require('../models/userModel');
+
 
 const getGoal = asyncHandler(async (req,res) =>{
-     const goals = await Goal.find();
+     const goals = await Goal.find({user : req.user.id});
     res.status(200).json(goals);
 })
 
@@ -15,7 +16,8 @@ const setGoal = asyncHandler(async (req,res) =>{
     }
     else{
         const goal= await Goal.create({
-            text : req.body.text
+            text : req.body.text,
+            user : req.user.id
         })
 
         res.status(200).json(goal);
@@ -26,7 +28,21 @@ const updateGoal = asyncHandler(async (req,res) =>{
     const goal = await Goal.findById(req.params.id)
     if(!goal){
         res.status(400)
-        throw new error("Goal Not found");
+        throw new Error("Goal Not found");
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //Check for User
+    if(!user){
+        res.status(400)
+        throw new Error("User not found")
+    }
+
+    //Make sure the logged user matches the  goal user
+    if(goal.user.toString() !== user.id){
+        res.status(400)
+        throw new Error("User not authorized")
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,{ new : true})
@@ -37,7 +53,21 @@ const deleteGoal = asyncHandler(async (req,res) =>{
     const goal = await Goal.findById(req.params.id)
     if(!goal){
         res.status(400)
-        throw new error ("Goal is not found")
+        throw new Error ("Goal is not found")
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //Check for User
+    if(!user){
+        res.status(400)
+        throw new Error("User not found")
+    }
+
+    //Make sure the logged user matches the  goal user
+    if(goal.user.toString() !== user.id){
+        res.status(400)
+        throw new Error("User not authorized")
     }
 
     const deletedGoal = await Goal.findByIdAndDelete (req.params.id)
